# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditLogListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AuditLogResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "count",
          "results"
        ],
        "title": "AuditLogListResponse",
        "type": "object"
      },
      "AuditLogResponse": {
        "properties": {
          "actorId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorid"
          },
          "actorType": {
            "enum": [
              "user",
              "system",
              "agent"
            ],
            "title": "Actortype",
            "type": "string"
          },
          "after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "caseId": {
            "title": "Caseid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "eventType": {
            "title": "Eventtype",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          }
        },
        "required": [
          "id",
          "caseId",
          "eventType",
          "actorType",
          "createdAt"
        ],
        "title": "AuditLogResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QaCaseListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QaCaseResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "count",
          "results",
          "limit",
          "offset"
        ],
        "title": "QaCaseListResponse",
        "type": "object"
      },
      "QaCaseResponse": {
        "properties": {
          "assignedTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignedto"
          },
          "className": {
            "title": "Classname",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "datasetId": {
            "title": "Datasetid",
            "type": "string"
          },
          "datasetVersion": {
            "title": "Datasetversion",
            "type": "string"
          },
          "errorType": {
            "title": "Errortype",
            "type": "string"
          },
          "evaluationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluationid"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "frameFileName": {
            "title": "Framefilename",
            "type": "string"
          },
          "frameIndex": {
            "title": "Frameindex",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "priority": {
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Priority",
            "type": "string"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "riskScore": {
            "title": "Riskscore",
            "type": "integer"
          },
          "sequenceId": {
            "title": "Sequenceid",
            "type": "string"
          },
          "sourceImageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceimageid"
          },
          "sourceSplit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcesplit"
          },
          "status": {
            "enum": [
              "unreviewed",
              "in_review",
              "confirmed",
              "corrected",
              "rejected",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          },
          "targetTrackId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targettrackid"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "datasetId",
          "datasetVersion",
          "sequenceId",
          "frameIndex",
          "frameFileName",
          "className",
          "errorType",
          "riskScore",
          "priority",
          "status",
          "evidence",
          "recommendation",
          "createdAt",
          "updatedAt"
        ],
        "title": "QaCaseResponse",
        "type": "object"
      },
      "QaCaseStatusRequest": {
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "in_review",
              "confirmed",
              "rejected",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "QaCaseStatusRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "QA orchestration API for 2D perception annotations",
    "title": "Label Guardian API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/qa-cases": {
      "get": {
        "operationId": "list_qa_cases_api_v1_qa_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unreviewed",
                    "in_review",
                    "confirmed",
                    "corrected",
                    "rejected",
                    "skipped"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sequenceId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sequenceid"
            }
          },
          {
            "in": "query",
            "name": "datasetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datasetid"
            }
          },
          {
            "in": "query",
            "name": "split",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Split"
            }
          },
          {
            "in": "query",
            "name": "sourceImageId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sourceimageid"
            }
          },
          {
            "in": "query",
            "name": "minRisk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minrisk"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Qa Cases",
        "tags": [
          "QA Cases"
        ]
      }
    },
    "/api/v1/qa-cases/{case_id}": {
      "get": {
        "operationId": "get_qa_case_api_v1_qa_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Qa Case",
        "tags": [
          "QA Cases"
        ]
      }
    },
    "/api/v1/qa-cases/{case_id}/audit": {
      "get": {
        "operationId": "get_qa_case_audit_api_v1_qa_cases__case_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Qa Case Audit",
        "tags": [
          "QA Cases"
        ]
      }
    },
    "/api/v1/qa-cases/{case_id}/status": {
      "post": {
        "operationId": "update_qa_case_status_api_v1_qa_cases__case_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QaCaseStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Qa Case Status",
        "tags": [
          "QA Cases"
        ]
      }
    },
...
  }
}